+++ integrations/sveltekit/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/sveltekit
 
+## 0.2.5
+
 ## 0.2.4
 
 ### Patch Changes